backend/pyproject.toml
[project]
name = "ml-stankin-proj"
version = "0.1.0"
description = ""
authors = [
    {name = "sadxyMath",email = "matvejzralko@gmail.com"}
]
readme = "README.md"
requires-python = "3.11.7"
dependencies = [
    "fastapi>=0.115.5,<0.116.0",
    "uvicorn[standard]>=0.32.0,<0.33.0",
    "transformers>=4.45.2,<4.46.0",
    "pandas>=2.2.3,<2.3.0",
    "pdfminer.six>=20231228,<20231229",
    "pydantic>=2.9.2,<2.10.0",
    "python-multipart>=0.0.12,<0.0.13",
    "python-docx>=1.1.2,<1.2.0"
]

[tool.poetry]
package-mode = false

[build-system]

requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
